refactor(shim): US-052 — découpe main.rs (2668 lignes) en modules
Code-motion pur. main.rs (orchestrateur: main/notify_session_end/
locate_sibling_hook_binary + tests) délègue à 3 modules :
- detect.rs : détection d'outil + localisation du vrai binaire
- exec.rs   : run_real + exit codes + signaux Unix (cfg-gated)
- hooks.rs  : config hooks Claude + Codex + tee JSONL Windows

Les cfg-gates plateforme sont désormais groupés et auditables. Imports
plateforme cfg-gated pour zéro warning sur chaque leg. Vérifié sur Linux
(clippy -D warnings) ET Windows (cargo check --target x86_64-pc-windows-msvc) ;
33 tests inchangés. (Codex+tee gardés dans hooks.rs car couplés à l'infra de
hooks Claude — fusion défendable vs 6 micro-modules à imports croisés.)
