+++ crates/paneflow-shim/src/exec.rs
@@ -1,0 +1,289 @@
+//! Run the real AI binary + Unix signal handling (US-052 split).
+
+use std::env;
+use std::ffi::OsString;
+use std::path::Path;
+use std::process::ExitCode;
+// The SIGINT watcher (`install_sigint_watcher` / `send_interrupt_stop`) is the
+// only consumer of these and is `#[cfg(unix)]`; gate the imports so the Windows
+// build (which compiles only `run_real` + `exit_code_from_status`) doesn't flag
+// them unused.
+#[cfg(unix)]
+use crate::locate_sibling_hook_binary;
+#[cfg(unix)]
+use std::io::Write;
+
+// ---------------------------------------------------------------------------
+// Run chain — spawn the real AI binary and wait for it
+// ---------------------------------------------------------------------------
+//
+
+// US-004 originally used `CommandExt::exec()` on Unix for zero-fork process
+// replacement. US-005 introduced the `HookConfigGuard` drop-cleanup contract,
+// which is incompatible with `exec()` — process replacement skips every Rust
+// destructor, so the guard would never fire. Both platforms now use
+// `Command::status()`; the shim pays one fork (~1-3 ms, well under the 15 ms
+// budget) in exchange for reliable cleanup.
+//
+// `Command` inherits the parent env by default, so `.envs(env::vars_os())`
+// is redundant — but the PRD AC bullet 5 lists it explicitly to make the
+// env-pass-through contract discoverable in the source. The `.env(...)`
+// calls afterward shadow per-key (Command::env is last-write-wins).
+//
+// PANEFLOW_AI_TOOL — set so `paneflow-ai-hook` (US-003) can tag every
+// outbound IPC frame with the right tool identity (`claude` vs `codex`).
+// Without this, `paneflow-ai-hook::detect_tool_from(None)` defaults to
+// `TOOL_CLAUDE`, which makes the sidebar render "Claude thinking…" for
+// every Codex turn — visible regression observed in the field.
+
+pub(crate) fn run_real(tool: &str, path: &Path, args: &[OsString]) -> ExitCode {
+    let mut cmd = std::process::Command::new(path);
+    cmd.args(args)
+        .envs(env::vars_os())
+        .env("PANEFLOW_AI_TOOL", tool)
+        // PANEFLOW_AI_PID — stable session identity propagated to every
+        // `paneflow-ai-hook` invocation fired by claude/codex during this
+        // session. Without it, the server's `Workspace::agent_sessions`
+        // (keyed by PID) collapses every Claude Code into one entry, so
+        // the sidebar shows `Claude thinking` for two concurrent sessions
+        // instead of `Claude thinking +1`. We use the shim's own PID
+        // (process::id()) rather than the child's because (a) the child
+        // PID isn't known until after spawn — too late for an env var on
+        // Command — and (b) the shim outlives the child via `waitpid`,
+        // so the PID stays reachable for the stale-PID sweep.
+        .env("PANEFLOW_AI_PID", std::process::id().to_string());
+
+    // Unix only: reset signal disposition + unblock SIGINT in the child.
+    //
+    // Required because Rust's `Command` inherits the parent's signal mask
+    // and dispositions across `execve`. The parent installs:
+    //   - `SIG_IGN` for SIGHUP/SIGTERM (shim survives PTY close / kill)
+    //   - `SIG_BLOCK` mask for SIGINT (consumed synchronously by the
+    //     `sigwait` thread in `install_sigint_watcher`, so the shim can
+    //     emit an `ai.stop` IPC frame on every Ctrl+C — including
+    //     mid-response interrupts where claude/codex intentionally fire
+    //     no `Stop` hook of their own).
+    //
+    // Without this `pre_exec` reset+unblock, the child would inherit both
+    // and Ctrl+C would do absolutely nothing (the AI would never see it,
+    // since `SIG_BLOCK`'d signals on a Linux process stay blocked across
+    // `execve`).
+    //
+    // `pre_exec` runs in the forked child between fork() and execve(). All
+    // calls below are async-signal-safe.
+    #[cfg(unix)]
+    unsafe {
+        use std::os::unix::process::CommandExt;
+        cmd.pre_exec(|| {
+            libc::signal(libc::SIGINT, libc::SIG_DFL);
+            libc::signal(libc::SIGHUP, libc::SIG_DFL);
+            libc::signal(libc::SIGTERM, libc::SIG_DFL);
+            let mut set: libc::sigset_t = std::mem::zeroed();
+            libc::sigemptyset(&mut set);
+            libc::sigaddset(&mut set, libc::SIGINT);
+            libc::sigaddset(&mut set, libc::SIGHUP);
+            libc::sigaddset(&mut set, libc::SIGTERM);
+            libc::pthread_sigmask(libc::SIG_UNBLOCK, &set, std::ptr::null_mut());
+
+            // US-016 (cli-hardening-followup-2026-Q3): on Linux,
+            // install PR_SET_PDEATHSIG so the spawned agent CLI
+            // (claude/codex/opencode) is killed by the kernel as
+            // soon as the shim's parent process (Paneflow) dies --
+            // even on a hard `kill -9` of Paneflow that bypasses
+            // any graceful Drop discipline. Without this, the agent
+            // is reparented to PID 1 and keeps streaming, burning
+            // the user's API tokens until its natural timeout.
+            //
+            // `parent_guard.rs` documents this gap on Linux/macOS
+            // because the GPUI app spawns through
+            // `portable-pty::CommandBuilder` which does not expose
+            // `pre_exec`. The shim wraps ~80% of those spawns
+            // (claude/codex/opencode all go through it) so this
+            // covers the realistic Linux path. macOS uses a stub
+            // pending kqueue NOTE_EXIT (out of US-016 scope).
+            //
+            // SAFETY: prctl is async-signal-safe; the call happens
+            // in the forked child between fork() and execve().
+            // A non-zero return is rare (only on a stripped kernel
+            // without prctl PR_SET_PDEATHSIG support) and is best-
+            // effort: we emit nothing on stderr because writing
+            // from pre_exec is not async-signal-safe in general,
+            // and we explicitly do NOT abort the exec -- letting
+            // the child run without PDEATHSIG is strictly better
+            // than failing the spawn entirely.
+            #[cfg(target_os = "linux")]
+            {
+                let _ = libc::prctl(
+                    libc::PR_SET_PDEATHSIG,
+                    libc::SIGKILL as libc::c_ulong,
+                    0,
+                    0,
+                    0,
+                );
+                // US-037: close the fork↔prctl race. PDEATHSIG only fires on
+                // a parent death that happens AFTER it's armed; if the shim
+                // (our parent) already died in the window between fork() and
+                // this prctl, the kernel never delivers the signal. getppid()
+                // == 1 means we were already reparented to init, so
+                // self-terminate rather than leaking a token-burning orphan.
+                // Both calls are async-signal-safe.
+                if libc::getppid() == 1 {
+                    libc::raise(libc::SIGKILL);
+                }
+            }
+            Ok(())
+        });
+    }
+
+    // Install signal isolation BEFORE spawn so the child inherits the
+    // mask/dispositions at fork (then `pre_exec` flips them back for the
+    // child only). Doing this BEFORE `cmd.spawn()` closes the race window
+    // where a Ctrl+C could land between spawn and signal-install.
+    #[cfg(unix)]
+    ignore_terminal_signals();
+    #[cfg(unix)]
+    install_sigint_watcher(tool);
+
+    let mut child = match cmd.spawn() {
+        Ok(c) => c,
+        Err(e) => {
+            eprintln!("paneflow-shim: spawn '{}' failed: {e}", path.display());
+            return ExitCode::from(127);
+        }
+    };
+
+    match child.wait() {
+        Ok(status) => exit_code_from_status(&status),
+        Err(e) => {
+            eprintln!("paneflow-shim: wait on '{}' failed: {e}", path.display());
+            ExitCode::from(1)
+        }
+    }
+}
+
+/// US-037: map a child `ExitStatus` to this process's `ExitCode`.
+///
+/// `status.code()` is `None` only when the child was terminated by a signal
+/// (Unix). The shell convention `128 + signum` (used by bash, `time(1)`, etc.)
+/// lets the parent terminal see the real cause (e.g. 130 for SIGINT, 139 for
+/// SIGSEGV) instead of an opaque `1`. Extracted to one place so the three
+/// `wait()` call sites stay consistent. `u8::try_from` clamps out-of-range
+/// codes to `1`.
+pub(crate) fn exit_code_from_status(status: &std::process::ExitStatus) -> ExitCode {
+    if let Some(code) = status.code() {
+        return ExitCode::from(u8::try_from(code).unwrap_or(1));
+    }
+    #[cfg(unix)]
+    {
+        use std::os::unix::process::ExitStatusExt;
+        if let Some(sig) = status.signal() {
+            let code = 128i32.saturating_add(sig);
+            return ExitCode::from(u8::try_from(code).unwrap_or(1));
+        }
+    }
+    ExitCode::from(1)
+}
+
+/// Make the shim survive PTY-close + kill signals so the child
+/// (claude/codex) can handle them without taking us down with it.
+///
+/// SIGINT is intentionally NOT in this list — it's handled by
+/// `install_sigint_watcher` via `sigwait` so we can emit a per-interrupt
+/// `ai.stop` IPC frame (mid-response Ctrl+C interrupts fire no hook from
+/// claude/codex, so this is the only signal we have).
+#[cfg(unix)]
+pub(crate) fn ignore_terminal_signals() {
+    // SAFETY: `libc::signal` with `SIG_IGN` is async-signal-safe and only
+    // mutates the kernel signal disposition table for the current process.
+    unsafe {
+        libc::signal(libc::SIGHUP, libc::SIG_IGN);
+        libc::signal(libc::SIGTERM, libc::SIG_IGN);
+    }
+}
+
+/// Block SIGINT in the shim, then spawn a dedicated thread that
+/// `sigwait`s on it. On every Ctrl+C, send `ai.stop` to PaneFlow so the
+/// sidebar loader transitions to `Finished` (then auto-resets to
+/// `Inactive` after 5s server-side). This is the ONLY way to detect a
+/// mid-stream interrupt because:
+///   - Claude Code does not fire its `Stop` hook when a turn is
+///     interrupted (only on natural completion).
+///   - Codex does not fire any hook on `esc`/Ctrl+C either.
+///
+/// `sigwait` is the POSIX-correct synchronous-from-thread receive: no
+/// async-signal-safety constraints, no self-pipe trick. Standard pattern
+/// (see Stevens APUE §12.8 "pthread_sigmask").
+#[cfg(unix)]
+pub(crate) fn install_sigint_watcher(tool: &str) {
+    // SAFETY: `pthread_sigmask` is thread-safe and only mutates the
+    // calling thread's signal mask. Blocking SIGINT here propagates to
+    // every thread spawned afterward (POSIX inheritance rule). The
+    // `pre_exec` hook in `run_real` re-unblocks SIGINT in the child.
+    unsafe {
+        let mut set: libc::sigset_t = std::mem::zeroed();
+        libc::sigemptyset(&mut set);
+        libc::sigaddset(&mut set, libc::SIGINT);
+        libc::pthread_sigmask(libc::SIG_BLOCK, &set, std::ptr::null_mut());
+    }
+
+    let tool = tool.to_owned();
+    let hook_path = locate_sibling_hook_binary();
+    std::thread::spawn(move || {
+        let Some(hook_path) = hook_path else {
+            return;
+        };
+        loop {
+            // SAFETY: `sigwait` blocks the calling thread until one of the
+            // signals in `set` is delivered to the process. Returns 0 on
+            // success and writes the received signal into `sig`. Spurious
+            // wakeups are not part of the POSIX contract; if it ever does
+            // return non-zero, exit the loop (the shim continues running,
+            // we just lose interrupt-driven notifications for this
+            // session — graceful degradation per PRD C4).
+            let sig = unsafe {
+                let mut set: libc::sigset_t = std::mem::zeroed();
+                libc::sigemptyset(&mut set);
+                libc::sigaddset(&mut set, libc::SIGINT);
+                let mut sig: libc::c_int = 0;
+                if libc::sigwait(&set, &mut sig) != 0 {
+                    return;
+                }
+                sig
+            };
+            if sig == libc::SIGINT {
+                send_interrupt_stop(&hook_path, &tool);
+            }
+        }
+    });
+}
+
+/// Spawn `paneflow-ai-hook Stop` with `{}` piped to stdin. Best-effort;
+/// any failure is silent (worst case: this Ctrl+C doesn't clear the
+/// loader, but the shim and the child remain unaffected).
+///
+/// Reaping policy: the wait happens on a detached helper thread, NOT on
+/// the calling sigwait thread. If the hook hangs (socket back-pressure,
+/// filesystem stall) the reaper thread hangs with it — but the sigwait
+/// thread stays responsive, so the next Ctrl+C lands as a fresh `ai.stop`
+/// rather than queuing behind the previous one. Without the helper, a
+/// dropped `Child` would leak a zombie until shim exit.
+#[cfg(unix)]
+pub(crate) fn send_interrupt_stop(hook_path: &Path, tool: &str) {
+    let Ok(mut child) = std::process::Command::new(hook_path)
+        .arg("Stop")
+        .env("PANEFLOW_AI_TOOL", tool)
+        .env("PANEFLOW_AI_PID", std::process::id().to_string())
+        .stdin(std::process::Stdio::piped())
+        .stdout(std::process::Stdio::null())
+        .stderr(std::process::Stdio::null())
+        .spawn()
+    else {
+        return;
+    };
+    if let Some(mut stdin) = child.stdin.take() {
+        let _ = stdin.write_all(b"{}");
+    }
+    std::thread::spawn(move || {
+        let _ = child.wait();
+    });
+}
