+++ crates/paneflow-shim/src/main.rs
@@ -30,1512 +30,16 @@
 
 use std::env;
 use std::ffi::OsString;
-use std::io::Write;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use std::process::ExitCode;
 
-// ---------------------------------------------------------------------------
-// Tool detection — from `current_exe()` filename stem
-// ---------------------------------------------------------------------------
-
-/// Read `std::env::current_exe()` and return the tool identity, or `None`
-/// if the shim was invoked under an unexpected name (direct `paneflow-shim`
-/// invocation, custom rename, etc.).
-fn detect_tool() -> Option<&'static str> {
-    let exe = env::current_exe().ok()?;
-    let stem = exe.file_stem()?.to_str()?;
-    detect_tool_from_stem(stem)
-}
-
-/// Testable inner: map a filename stem to the tool identity. Only the two
-/// exact lowercase matches are accepted — US-008 controls the extracted
-/// filenames, so anything else here means the binary has been renamed or
-/// invoked directly.
-fn detect_tool_from_stem(stem: &str) -> Option<&'static str> {
-    match stem {
-        "claude" => Some("claude"),
-        "codex" => Some("codex"),
-        _ => None,
-    }
-}
-
-// ---------------------------------------------------------------------------
-// PATH walk — find the real AI binary, excluding the shim's own directory
-// ---------------------------------------------------------------------------
-
-/// Candidate executable names to probe in each `$PATH` entry. Unix looks for
-/// a bare filename; Windows tries `.exe` first, then `.cmd` (covers both
-/// native AI builds and Node-shipped wrappers like `claude.cmd`).
-#[cfg(unix)]
-fn candidate_names(tool: &str) -> Vec<String> {
-    vec![tool.to_owned()]
-}
-
-#[cfg(windows)]
-fn candidate_names(tool: &str) -> Vec<String> {
-    vec![format!("{tool}.exe"), format!("{tool}.cmd")]
-}
-
-/// Walk `$PATH` and return the first entry that contains a matching
-/// executable, skipping the shim's own directory AND any candidate
-/// that is the shim binary itself by inode (US-017 hardlink defense).
-fn find_real_binary(tool: &str) -> Option<PathBuf> {
-    let path_var = env::var_os("PATH")?;
-    // Per `install_method.rs:92-98`, always canonicalize `current_exe()` —
-    // on Linux it may point at `/proc/self/exe` or follow through a symlink.
-    let self_exe = env::current_exe().ok();
-    let self_dir = self_exe
-        .as_deref()
-        .and_then(|p| p.parent().map(Path::to_path_buf));
-
-    find_real_binary_in(
-        tool,
-        env::split_paths(&path_var),
-        self_dir.as_deref(),
-        self_exe.as_deref(),
-    )
-}
-
-/// Pure inner — takes PATH entries as an iterator and optional
-/// self-dir / self-exe paths, so tests can pass a controlled set
-/// without mutating `$PATH` or relying on `current_exe()`.
-fn find_real_binary_in<I>(
-    tool: &str,
-    path_entries: I,
-    self_dir: Option<&Path>,
-    self_exe: Option<&Path>,
-) -> Option<PathBuf>
-where
-    I: IntoIterator<Item = PathBuf>,
-{
-    // Canonicalize once; `None` if the dir doesn't exist (in which case we
-    // can't match anything against it, so the self-exclusion is a no-op and
-    // PATH is walked in full — safer than silently skipping nothing).
-    let self_canon = self_dir.and_then(|d| std::fs::canonicalize(d).ok());
-    // US-017 (cli-hardening-followup-2026-Q3): capture the shim's
-    // own file identity. A candidate matching
-    // this identity is the shim itself reached via a hardlink and
-    // must be skipped to break the recursive-spawn loop.
-    let self_identity = self_exe.and_then(file_identity);
-    let candidates = candidate_names(tool);
-
-    for dir in path_entries {
-        if same_canonical_dir(&self_canon, &dir) {
-            continue;
-        }
-        for name in &candidates {
-            let candidate = dir.join(name);
-            if !candidate.is_file() {
-                continue;
-            }
-            // US-037: on Unix, require the executable bit too. A non-executable
-            // homonym (e.g. a `0644` data file named like the tool) earlier in
-            // `$PATH` would otherwise be returned, and the subsequent spawn
-            // fails `EACCES`/`ENOEXEC` *without* continuing the walk (unlike
-            // `execvp`, which skips it). Skip it here so the real binary later
-            // in `$PATH` is found.
-            #[cfg(unix)]
-            {
-                use std::os::unix::fs::PermissionsExt;
-                let executable = std::fs::metadata(&candidate)
-                    .map(|m| m.permissions().mode() & 0o111 != 0)
-                    .unwrap_or(false);
-                if !executable {
-                    continue;
-                }
-            }
-            if is_same_file_as_shim(&self_identity, &candidate) {
-                // US-017: hardlink-loop guard. The shim has no `log`
-                // dep (would bloat the small binary); stderr is
-                // already the diagnostic channel used elsewhere
-                // in this binary (cf. line ~287).
-                eprintln!(
-                    "paneflow-shim: skipping {} -- matches shim identity (hardlink loop guard)",
-                    candidate.display()
-                );
-                continue;
-            }
-            return Some(candidate);
-        }
-    }
-    None
-}
-
-/// US-017 (cli-hardening-followup-2026-Q3): capture a file's
-/// cross-platform identity. `same_file::Handle` uses the native
-/// file identity on Unix and Windows, so hardlinks compare equal
-/// without relying on unstable standard-library APIs.
-///
-/// Returns `None` when the path cannot be opened, in which case
-/// the comparison degrades to a no-op and the dir-canonicalize
-/// path remains the residual defense.
-fn file_identity(path: &Path) -> Option<same_file::Handle> {
-    same_file::Handle::from_path(path).ok()
-}
-
-fn is_same_file_as_shim(self_identity: &Option<same_file::Handle>, candidate: &Path) -> bool {
-    match (self_identity.as_ref(), file_identity(candidate)) {
-        (Some(a), Some(b)) => a == &b,
-        _ => false,
-    }
-}
-
-/// Returns `true` when `dir` canonicalizes to the same path as `self_canon`.
-/// Handles symlinks, trailing slashes, and `..` segments that would otherwise
-/// make two string-equal paths compare as different and vice versa.
-///
-/// This stays as the cheap fast-path filter that skips the shim's
-/// own install directory wholesale (saves one `metadata` syscall
-/// per candidate when the dir matches by canonicalization). The
-/// inode-level defense against a hardlinked shim planted in a
-/// different `$PATH` directory lives in [`is_same_file_as_shim`]
-/// and runs for every candidate that survives the dir filter
-/// (US-017 cli-hardening-followup-2026-Q3).
-fn same_canonical_dir(self_canon: &Option<PathBuf>, dir: &Path) -> bool {
-    match (self_canon, std::fs::canonicalize(dir).ok()) {
-        (Some(s), Some(d)) => *s == d,
-        _ => false,
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Run chain — spawn the real AI binary and wait for it
-// ---------------------------------------------------------------------------
-//
-// US-004 originally used `CommandExt::exec()` on Unix for zero-fork process
-// replacement. US-005 introduced the `HookConfigGuard` drop-cleanup contract,
-// which is incompatible with `exec()` — process replacement skips every Rust
-// destructor, so the guard would never fire. Both platforms now use
-// `Command::status()`; the shim pays one fork (~1-3 ms, well under the 15 ms
-// budget) in exchange for reliable cleanup.
-//
-// `Command` inherits the parent env by default, so `.envs(env::vars_os())`
-// is redundant — but the PRD AC bullet 5 lists it explicitly to make the
-// env-pass-through contract discoverable in the source. The `.env(...)`
-// calls afterward shadow per-key (Command::env is last-write-wins).
-//
-// PANEFLOW_AI_TOOL — set so `paneflow-ai-hook` (US-003) can tag every
-// outbound IPC frame with the right tool identity (`claude` vs `codex`).
-// Without this, `paneflow-ai-hook::detect_tool_from(None)` defaults to
-// `TOOL_CLAUDE`, which makes the sidebar render "Claude thinking…" for
-// every Codex turn — visible regression observed in the field.
-
-fn run_real(tool: &str, path: &Path, args: &[OsString]) -> ExitCode {
-    let mut cmd = std::process::Command::new(path);
-    cmd.args(args)
-        .envs(env::vars_os())
-        .env("PANEFLOW_AI_TOOL", tool)
-        // PANEFLOW_AI_PID — stable session identity propagated to every
-        // `paneflow-ai-hook` invocation fired by claude/codex during this
-        // session. Without it, the server's `Workspace::agent_sessions`
-        // (keyed by PID) collapses every Claude Code into one entry, so
-        // the sidebar shows `Claude thinking` for two concurrent sessions
-        // instead of `Claude thinking +1`. We use the shim's own PID
-        // (process::id()) rather than the child's because (a) the child
-        // PID isn't known until after spawn — too late for an env var on
-        // Command — and (b) the shim outlives the child via `waitpid`,
-        // so the PID stays reachable for the stale-PID sweep.
-        .env("PANEFLOW_AI_PID", std::process::id().to_string());
-
-    // Unix only: reset signal disposition + unblock SIGINT in the child.
-    //
-    // Required because Rust's `Command` inherits the parent's signal mask
-    // and dispositions across `execve`. The parent installs:
-    //   - `SIG_IGN` for SIGHUP/SIGTERM (shim survives PTY close / kill)
-    //   - `SIG_BLOCK` mask for SIGINT (consumed synchronously by the
-    //     `sigwait` thread in `install_sigint_watcher`, so the shim can
-    //     emit an `ai.stop` IPC frame on every Ctrl+C — including
-    //     mid-response interrupts where claude/codex intentionally fire
-    //     no `Stop` hook of their own).
-    //
-    // Without this `pre_exec` reset+unblock, the child would inherit both
-    // and Ctrl+C would do absolutely nothing (the AI would never see it,
-    // since `SIG_BLOCK`'d signals on a Linux process stay blocked across
-    // `execve`).
-    //
-    // `pre_exec` runs in the forked child between fork() and execve(). All
-    // calls below are async-signal-safe.
-    #[cfg(unix)]
-    unsafe {
-        use std::os::unix::process::CommandExt;
-        cmd.pre_exec(|| {
-            libc::signal(libc::SIGINT, libc::SIG_DFL);
-            libc::signal(libc::SIGHUP, libc::SIG_DFL);
-            libc::signal(libc::SIGTERM, libc::SIG_DFL);
-            let mut set: libc::sigset_t = std::mem::zeroed();
-            libc::sigemptyset(&mut set);
-            libc::sigaddset(&mut set, libc::SIGINT);
-            libc::sigaddset(&mut set, libc::SIGHUP);
-            libc::sigaddset(&mut set, libc::SIGTERM);
-            libc::pthread_sigmask(libc::SIG_UNBLOCK, &set, std::ptr::null_mut());
-
-            // US-016 (cli-hardening-followup-2026-Q3): on Linux,
-            // install PR_SET_PDEATHSIG so the spawned agent CLI
-            // (claude/codex/opencode) is killed by the kernel as
-            // soon as the shim's parent process (Paneflow) dies --
-            // even on a hard `kill -9` of Paneflow that bypasses
-            // any graceful Drop discipline. Without this, the agent
-            // is reparented to PID 1 and keeps streaming, burning
-            // the user's API tokens until its natural timeout.
-            //
-            // `parent_guard.rs` documents this gap on Linux/macOS
-            // because the GPUI app spawns through
-            // `portable-pty::CommandBuilder` which does not expose
-            // `pre_exec`. The shim wraps ~80% of those spawns
-            // (claude/codex/opencode all go through it) so this
-            // covers the realistic Linux path. macOS uses a stub
-            // pending kqueue NOTE_EXIT (out of US-016 scope).
-            //
-            // SAFETY: prctl is async-signal-safe; the call happens
-            // in the forked child between fork() and execve().
-            // A non-zero return is rare (only on a stripped kernel
-            // without prctl PR_SET_PDEATHSIG support) and is best-
-            // effort: we emit nothing on stderr because writing
-            // from pre_exec is not async-signal-safe in general,
-            // and we explicitly do NOT abort the exec -- letting
-            // the child run without PDEATHSIG is strictly better
-            // than failing the spawn entirely.
-            #[cfg(target_os = "linux")]
-            {
-                let _ = libc::prctl(
-                    libc::PR_SET_PDEATHSIG,
-                    libc::SIGKILL as libc::c_ulong,
-                    0,
-                    0,
-                    0,
-                );
-                // US-037: close the fork↔prctl race. PDEATHSIG only fires on
-                // a parent death that happens AFTER it's armed; if the shim
-                // (our parent) already died in the window between fork() and
-                // this prctl, the kernel never delivers the signal. getppid()
-                // == 1 means we were already reparented to init, so
-                // self-terminate rather than leaking a token-burning orphan.
-                // Both calls are async-signal-safe.
-                if libc::getppid() == 1 {
-                    libc::raise(libc::SIGKILL);
-                }
-            }
-            Ok(())
-        });
-    }
-
-    // Install signal isolation BEFORE spawn so the child inherits the
-    // mask/dispositions at fork (then `pre_exec` flips them back for the
-    // child only). Doing this BEFORE `cmd.spawn()` closes the race window
-    // where a Ctrl+C could land between spawn and signal-install.
-    #[cfg(unix)]
-    ignore_terminal_signals();
-    #[cfg(unix)]
-    install_sigint_watcher(tool);
-
-    let mut child = match cmd.spawn() {
-        Ok(c) => c,
-        Err(e) => {
-            eprintln!("paneflow-shim: spawn '{}' failed: {e}", path.display());
-            return ExitCode::from(127);
-        }
-    };
-
-    match child.wait() {
-        Ok(status) => exit_code_from_status(&status),
-        Err(e) => {
-            eprintln!("paneflow-shim: wait on '{}' failed: {e}", path.display());
-            ExitCode::from(1)
-        }
-    }
-}
-
-/// US-037: map a child `ExitStatus` to this process's `ExitCode`.
-///
-/// `status.code()` is `None` only when the child was terminated by a signal
-/// (Unix). The shell convention `128 + signum` (used by bash, `time(1)`, etc.)
-/// lets the parent terminal see the real cause (e.g. 130 for SIGINT, 139 for
-/// SIGSEGV) instead of an opaque `1`. Extracted to one place so the three
-/// `wait()` call sites stay consistent. `u8::try_from` clamps out-of-range
-/// codes to `1`.
-fn exit_code_from_status(status: &std::process::ExitStatus) -> ExitCode {
-    if let Some(code) = status.code() {
-        return ExitCode::from(u8::try_from(code).unwrap_or(1));
-    }
-    #[cfg(unix)]
-    {
-        use std::os::unix::process::ExitStatusExt;
-        if let Some(sig) = status.signal() {
-            let code = 128i32.saturating_add(sig);
-            return ExitCode::from(u8::try_from(code).unwrap_or(1));
-        }
-    }
-    ExitCode::from(1)
-}
-
-/// Make the shim survive PTY-close + kill signals so the child
-/// (claude/codex) can handle them without taking us down with it.
-///
-/// SIGINT is intentionally NOT in this list — it's handled by
-/// `install_sigint_watcher` via `sigwait` so we can emit a per-interrupt
-/// `ai.stop` IPC frame (mid-response Ctrl+C interrupts fire no hook from
-/// claude/codex, so this is the only signal we have).
-#[cfg(unix)]
-fn ignore_terminal_signals() {
-    // SAFETY: `libc::signal` with `SIG_IGN` is async-signal-safe and only
-    // mutates the kernel signal disposition table for the current process.
-    unsafe {
-        libc::signal(libc::SIGHUP, libc::SIG_IGN);
-        libc::signal(libc::SIGTERM, libc::SIG_IGN);
-    }
-}
-
-/// Block SIGINT in the shim, then spawn a dedicated thread that
-/// `sigwait`s on it. On every Ctrl+C, send `ai.stop` to PaneFlow so the
-/// sidebar loader transitions to `Finished` (then auto-resets to
-/// `Inactive` after 5s server-side). This is the ONLY way to detect a
-/// mid-stream interrupt because:
-///   - Claude Code does not fire its `Stop` hook when a turn is
-///     interrupted (only on natural completion).
-///   - Codex does not fire any hook on `esc`/Ctrl+C either.
-///
-/// `sigwait` is the POSIX-correct synchronous-from-thread receive: no
-/// async-signal-safety constraints, no self-pipe trick. Standard pattern
-/// (see Stevens APUE §12.8 "pthread_sigmask").
-#[cfg(unix)]
-fn install_sigint_watcher(tool: &str) {
-    // SAFETY: `pthread_sigmask` is thread-safe and only mutates the
-    // calling thread's signal mask. Blocking SIGINT here propagates to
-    // every thread spawned afterward (POSIX inheritance rule). The
-    // `pre_exec` hook in `run_real` re-unblocks SIGINT in the child.
-    unsafe {
-        let mut set: libc::sigset_t = std::mem::zeroed();
-        libc::sigemptyset(&mut set);
-        libc::sigaddset(&mut set, libc::SIGINT);
-        libc::pthread_sigmask(libc::SIG_BLOCK, &set, std::ptr::null_mut());
-    }
-
-    let tool = tool.to_owned();
-    let hook_path = locate_sibling_hook_binary();
-    std::thread::spawn(move || {
-        let Some(hook_path) = hook_path else {
-            return;
-        };
-        loop {
-            // SAFETY: `sigwait` blocks the calling thread until one of the
-            // signals in `set` is delivered to the process. Returns 0 on
-            // success and writes the received signal into `sig`. Spurious
-            // wakeups are not part of the POSIX contract; if it ever does
-            // return non-zero, exit the loop (the shim continues running,
-            // we just lose interrupt-driven notifications for this
-            // session — graceful degradation per PRD C4).
-            let sig = unsafe {
-                let mut set: libc::sigset_t = std::mem::zeroed();
-                libc::sigemptyset(&mut set);
-                libc::sigaddset(&mut set, libc::SIGINT);
-                let mut sig: libc::c_int = 0;
-                if libc::sigwait(&set, &mut sig) != 0 {
-                    return;
-                }
-                sig
-            };
-            if sig == libc::SIGINT {
-                send_interrupt_stop(&hook_path, &tool);
-            }
-        }
-    });
-}
-
-/// Spawn `paneflow-ai-hook Stop` with `{}` piped to stdin. Best-effort;
-/// any failure is silent (worst case: this Ctrl+C doesn't clear the
-/// loader, but the shim and the child remain unaffected).
-///
-/// Reaping policy: the wait happens on a detached helper thread, NOT on
-/// the calling sigwait thread. If the hook hangs (socket back-pressure,
-/// filesystem stall) the reaper thread hangs with it — but the sigwait
-/// thread stays responsive, so the next Ctrl+C lands as a fresh `ai.stop`
-/// rather than queuing behind the previous one. Without the helper, a
-/// dropped `Child` would leak a zombie until shim exit.
-#[cfg(unix)]
-fn send_interrupt_stop(hook_path: &Path, tool: &str) {
-    use std::io::Write;
-    let Ok(mut child) = std::process::Command::new(hook_path)
-        .arg("Stop")
-        .env("PANEFLOW_AI_TOOL", tool)
-        .env("PANEFLOW_AI_PID", std::process::id().to_string())
-        .stdin(std::process::Stdio::piped())
-        .stdout(std::process::Stdio::null())
-        .stderr(std::process::Stdio::null())
-        .spawn()
-    else {
-        return;
-    };
-    if let Some(mut stdin) = child.stdin.take() {
-        let _ = stdin.write_all(b"{}");
-    }
-    std::thread::spawn(move || {
-        let _ = child.wait();
-    });
-}
-
-// ---------------------------------------------------------------------------
-// Hook config injection (US-005) — idempotent `.claude/settings.local.json`
-// ---------------------------------------------------------------------------
-
-/// Claude Code 2.x hook events the shim registers handlers for. `SubagentStop`
-/// is intentionally omitted — the server maps it to `ai.stop` identically to
-/// `Stop` (US-002), and registering both would produce duplicate IPC frames.
-const CLAUDE_HOOK_EVENTS: &[&str] = &[
-    "UserPromptSubmit",
-    "Notification",
-    "Stop",
-    "PreToolUse",
-    "PostToolUse",
-];
-
-/// Bare-name fallback used when `locate_sibling_hook_binary()` cannot resolve
-/// the absolute path to the sibling hook binary (test harness, exotic FS,
-/// `current_exe` failure). Detection (`is_paneflow_hook_command`) is
-/// basename-based so this fallback is also recognized for cleanup, alongside
-/// the absolute-path form normally written by `resolve_hook_command`.
-///
-/// Cleanup matches by basename as a belt-and-suspenders complement to the
-/// `_paneflow_managed` marker: Claude Code does not guarantee unknown fields
-/// survive re-serialization (anthropics/claude-code#5886), so the command
-/// string is the only round-trip-stable identifier we can rely on.
-///
-/// Known namespace-collision limitation: any user-authored hook command whose
-/// program basename is literally `paneflow-ai-hook` (or `.exe` on Windows)
-/// will be treated as PaneFlow-managed and removed on cleanup. The basename
-/// rule narrows this further than the previous bare-prefix rule did, but the
-/// theoretical collision remains.
-const HOOK_COMMAND_PREFIX: &str = "paneflow-ai-hook ";
-
-/// Render `path` for inclusion in an `eprintln!` going to the user's
-/// terminal. Replaces bytes outside the printable ASCII range (`0x20..=0x7E`)
-/// with `?` to defuse ANSI-escape-sequence injection via a maliciously
-/// named CWD or `.claude/` directory. Path content is never a secret — it
-/// was always going to be visible on stderr — but without this scrub, a
-/// crafted directory could clear the screen, set the terminal title, or
-/// inject false log lines (Phase 7 security audit, MEDIUM finding).
-fn safe_path_display(path: &Path) -> String {
-    path.display()
-        .to_string()
-        .chars()
-        .map(|c| if (' '..='~').contains(&c) { c } else { '?' })
-        .collect()
-}
-
-/// Returns true iff `$PANEFLOW_SOCKET_PATH` is set and points at an existing
-/// filesystem entry. We treat unset/unreachable as "PaneFlow not running":
-/// in that state, installing hook config is pointless (the hooks would
-/// invoke `paneflow-ai-hook`, which would fail silently per PRD constraint
-/// C4) and we instead sweep any orphan entries left by a previous SIGKILL'd
-/// session. Existence-only check keeps this cross-platform — Unix sockets
-/// and Windows named pipes both surface as `Path::exists() == true` when
-/// the listener is bound.
-fn paneflow_ipc_reachable() -> bool {
-    let Some(raw) = env::var_os("PANEFLOW_SOCKET_PATH") else {
-        return false;
-    };
-    if raw.is_empty() {
-        return false;
-    }
-    Path::new(&raw).exists()
-}
-
-/// Best-effort removal of PaneFlow-managed entries from an existing hook
-/// config file when no active IPC channel is reachable. Reads, runs
-/// `remove_fn`, writes back (or deletes if the file is now empty). All
-/// failures swallow silently — a sweep that fails just retries on the
-/// next shim invocation. Used by `install()` to clean up after a previous
-/// SIGKILL'd session that never got to fire its `Drop` impl.
-fn sweep_orphan_hook_config(settings_path: &Path, remove_fn: fn(&mut serde_json::Value)) {
-    let Ok(content) = std::fs::read_to_string(settings_path) else {
-        return;
-    };
-    let Ok(mut root) = serde_json::from_str::<serde_json::Value>(&content) else {
-        return;
-    };
-    let before = root.clone();
-    remove_fn(&mut root);
-    if root == before {
-        // Nothing to sweep — file has no PaneFlow entries.
-        return;
-    }
-    let is_empty = root
-        .as_object()
-        .map(serde_json::Map::is_empty)
-        .unwrap_or(false);
-    let _ = if is_empty {
-        std::fs::remove_file(settings_path)
-    } else {
-        write_atomic(settings_path, &root)
-    };
-}
-
-/// Shared scaffold for both Claude and Codex hook config installation:
-/// validate the config dir (creating it if absent), read+parse any existing
-/// JSON tree (treating corrupt JSON as empty), apply `merge_fn`, and write
-/// atomically. Returns `Some((settings_path, created_dir))` on success;
-/// `None` when the filesystem refuses our writes or the config dir is
-/// occupied by a non-directory.
-///
-/// Both `HookConfigGuard::install_at` (Claude) and
-/// `CodexHookConfigGuard::install_at` (Codex) layer their type-construction
-/// on top of this — the heavy filesystem + JSON work was previously
-/// duplicated across ~80 lines apart.
-fn install_hook_config_file(
-    config_dir: &Path,
-    config_filename: &str,
-    tool_label: &str,
-    merge_fn: fn(&mut serde_json::Value),
-) -> Option<(PathBuf, bool)> {
-    let settings_path = config_dir.join(config_filename);
-    // `exists()` returns true for both files and directories; we need to
-    // distinguish so a stale `.claude`/`.codex` regular file (e.g. left
-    // behind by an earlier tool) doesn't masquerade as a usable directory
-    // and silently break hook injection further down. Surface the case
-    // with an actionable message instead of letting `write_atomic` fail
-    // with a cryptic ENOTDIR from inside `tempfile::NamedTempFile::new_in`.
-    let existed_as_dir = config_dir.is_dir();
-    let exists_as_other = config_dir.exists() && !existed_as_dir;
-
-    if exists_as_other {
-        eprintln!(
-            "paneflow-shim: {} exists but is not a directory; remove or \
-             rename it to enable {tool_label} hooks this session",
-            safe_path_display(config_dir)
-        );
-        return None;
-    }
-
-    if !existed_as_dir {
-        if let Err(e) = std::fs::create_dir_all(config_dir) {
-            eprintln!(
-                "paneflow-shim: cannot create {} ({e}); {tool_label} hooks \
-                 disabled this session",
-                safe_path_display(config_dir)
-            );
-            return None;
-        }
-    }
-
-    let existing = std::fs::read_to_string(&settings_path).unwrap_or_default();
-    let mut root: serde_json::Value = if existing.trim().is_empty() {
-        serde_json::json!({})
-    } else {
-        match serde_json::from_str(&existing) {
-            Ok(v) => v,
-            Err(e) => {
-                // Corrupt JSON treated as empty; overwriting is preferable
-                // to aborting the shim and leaving the user with a broken
-                // settings file they can't fix from inside the AI tool.
-                eprintln!(
-                    "paneflow-shim: {} contained invalid JSON ({e}); \
-                     overwriting with a fresh config",
-                    safe_path_display(&settings_path)
-                );
-                serde_json::json!({})
-            }
-        }
-    };
-
-    merge_fn(&mut root);
-
-    if let Err(e) = write_atomic(&settings_path, &root) {
-        eprintln!(
-            "paneflow-shim: cannot write {} ({e}); {tool_label} hooks \
-             disabled this session",
-            safe_path_display(&settings_path)
-        );
-        if !existed_as_dir {
-            let _ = std::fs::remove_dir(config_dir);
-        }
-        return None;
-    }
-
-    Some((settings_path, !existed_as_dir))
-}
-
-/// Shared cleanup used by both guards' `Drop` impls: read the settings
-/// file, run `remove_fn` to strip PaneFlow's entries, then either delete
-/// the file (if now empty) and rmdir the config dir (if we created it),
-/// or write the cleaned tree back atomically. All failures swallow
-/// silently — Drop must never panic, and any error here means the next
-/// shim invocation's merge-idempotency will converge the state.
-fn cleanup_hook_config_file(
-    settings_path: &Path,
-    config_dir: &Path,
-    created_dir: bool,
-    remove_fn: fn(&mut serde_json::Value),
-) {
-    let Ok(content) = std::fs::read_to_string(settings_path) else {
-        return;
-    };
-    let Ok(mut root) = serde_json::from_str::<serde_json::Value>(&content) else {
-        return;
-    };
-
-    remove_fn(&mut root);
-
-    let is_empty = root
-        .as_object()
-        .map(serde_json::Map::is_empty)
-        .unwrap_or(false);
-
-    if is_empty {
-        let _ = std::fs::remove_file(settings_path);
-        if created_dir {
-            // `remove_dir` only succeeds if the directory is empty — safe
-            // even if the user dropped other files into the config dir.
-            let _ = std::fs::remove_dir(config_dir);
-        }
-    } else {
-        let _ = write_atomic(settings_path, &root);
-    }
-}
-
-/// RAII guard: writes PaneFlow's hook config on construction, removes it on
-/// drop. The guard must live for the duration of the child Claude Code
-/// process, then drop normally when `main()` returns — this is why US-005
-/// forces `run_real()` (vs `exec()`) so destructors actually fire.
-struct HookConfigGuard {
-    settings_path: PathBuf,
-    claude_dir: PathBuf,
-    // Whether the shim created `.claude/`. Only rmdir if we created it, so we
-    // don't clobber a user-created directory that happened to be empty.
-    created_dir: bool,
-}
-
-impl HookConfigGuard {
-    /// Install in the project CWD. Returns `None` if the filesystem refuses
-    /// our writes (read-only, permission denied, etc.) — the shim proceeds
-    /// without hooks in that case, per PRD constraint C4. Also returns
-    /// `None` (after sweeping any orphan entries left by a previous
-    /// SIGKILL'd session) when no PaneFlow IPC socket is reachable: writing
-    /// hook config that would invoke a dead handler would just create
-    /// config noise per C4.
-    fn install() -> Option<Self> {
-        let cwd = env::current_dir().ok()?;
-        let claude_dir = cwd.join(".claude");
-        if !paneflow_ipc_reachable() {
-            sweep_orphan_hook_config(
-                &claude_dir.join("settings.local.json"),
-                remove_paneflow_hooks,
-            );
-            return None;
-        }
-        Self::install_at(&claude_dir)
-    }
-
-    /// Testable inner. Takes the absolute path to the `.claude/` directory.
-    /// Does NOT check `PANEFLOW_SOCKET_PATH` — the orphan-sweep gate lives
-    /// in `install()` so unit tests can drive `install_at` without
-    /// fabricating a live IPC socket.
-    fn install_at(claude_dir: &Path) -> Option<Self> {
-        let (settings_path, created_dir) = install_hook_config_file(
-            claude_dir,
-            "settings.local.json",
-            "Claude Code",
-            merge_paneflow_hooks,
-        )?;
-        Some(Self {
-            settings_path,
-            claude_dir: claude_dir.to_path_buf(),
-            created_dir,
-        })
-    }
-}
-
-impl Drop for HookConfigGuard {
-    fn drop(&mut self) {
-        cleanup_hook_config_file(
-            &self.settings_path,
-            &self.claude_dir,
-            self.created_dir,
-            remove_paneflow_hooks,
-        );
-    }
-}
-
-/// Build the `"command"` string written into `settings.local.json` /
-/// `hooks.json` for the given hook `event`. Prefers the absolute path to the
-/// sibling `paneflow-ai-hook` binary (resolved via `current_exe().parent()`)
-/// so hooks resolve even when Claude Code or Codex is launched outside the
-/// PATH-injected shell paneflow normally provides — e.g., when a user runs
-/// `claude` directly in a project where a previous shim invocation left a
-/// managed `settings.local.json` in place. Falls back to the bare binary name
-/// when `current_exe()` fails or the sibling isn't present (test harnesses
-/// where the test binary lives in `target/debug/deps/`, exotic filesystems);
-/// the bare form is still recognized by `is_paneflow_hook_command` for
-/// detection and cleanup.
-fn resolve_hook_command(event: &str) -> String {
-    match locate_sibling_hook_binary() {
-        Some(path) => format!("{} {}", path.display(), event),
-        None => format!("{HOOK_COMMAND_PREFIX}{event}"),
-    }
-}
-
-/// Returns `true` if `command` is a paneflow-managed hook command, regardless
-/// of whether it uses the legacy bare-name format (`paneflow-ai-hook <Event>`)
-/// or the absolute-path format produced by `resolve_hook_command`. Detection
-/// is basename-based so it works across both shapes and across platforms
-/// (`paneflow-ai-hook` on Unix, `paneflow-ai-hook.exe` on Windows).
-///
-/// Intentionally does NOT verify that the binary exists on disk: a stale
-/// config pointing at a removed cache dir (e.g., user uninstalled paneflow,
-/// `cargo clean` between sessions) must still be recognized so cleanup can
-/// remove it on the next shim run.
-fn is_paneflow_hook_command(command: &str) -> bool {
-    let first_token = command.split_whitespace().next().unwrap_or("");
-    let basename = Path::new(first_token)
-        .file_name()
-        .and_then(|s| s.to_str())
-        .unwrap_or(first_token);
-    basename == "paneflow-ai-hook" || basename == "paneflow-ai-hook.exe"
-}
-
-/// Merge PaneFlow's hook handlers into the parsed settings tree. Idempotent:
-/// if a PaneFlow handler for an event is already present (identified by the
-/// command basename OR the `_paneflow_managed` marker), we don't duplicate.
-fn merge_paneflow_hooks(root: &mut serde_json::Value) {
-    let root_obj = match root.as_object_mut() {
-        Some(o) => o,
-        None => {
-            *root = serde_json::json!({});
-            // re-borrow after replacement — `if let Some(...)` because we
-            // just assigned an object, so this unwrap is infallible.
-            let Some(o) = root.as_object_mut() else {
-                return;
-            };
-            o
-        }
-    };
-
-    let hooks_entry = root_obj
-        .entry("hooks")
-        .or_insert_with(|| serde_json::json!({}));
-    let Some(hooks_obj) = hooks_entry.as_object_mut() else {
-        // User's `hooks` key is not an object (e.g., user set it to a
-        // string by mistake). Overwrite it — we own the managed entries.
-        *hooks_entry = serde_json::json!({});
-        return;
-    };
-
-    for event in CLAUDE_HOOK_EVENTS {
-        let entry = hooks_obj
-            .entry(*event)
-            .or_insert_with(|| serde_json::json!([]));
-        let Some(array) = entry.as_array_mut() else {
-            // User's event value is not an array; skip this event rather
-            // than clobber what might be intentional config.
-            continue;
-        };
-
-        let already_installed = array.iter().any(is_paneflow_matcher_group);
-        if already_installed {
-            continue;
-        }
-
-        // The `_paneflow_managed` marker sits on the OUTER matcher-group
-        // wrapper — that's where `is_paneflow_matcher_group` checks it.
-        // The inner handler object carries only the Claude-Code-native
-        // fields so we don't send unexpected custom fields to Claude Code's
-        // command runner. Identification falls back to the `command`
-        // basename if the outer marker is stripped.
-        array.push(serde_json::json!({
-            "_paneflow_managed": true,
-            "hooks": [
-                {
-                    "type": "command",
-                    "command": resolve_hook_command(event),
-                    "timeout": 5,
-                }
-            ]
-        }));
-    }
-}
-
-/// Remove PaneFlow's hook handlers from the parsed settings tree. Leaves
-/// user entries untouched. Collapses empty event arrays and the empty
-/// `hooks` key so cleanup produces a minimal file.
-fn remove_paneflow_hooks(root: &mut serde_json::Value) {
-    let Some(root_obj) = root.as_object_mut() else {
-        return;
-    };
-    let Some(hooks_obj) = root_obj.get_mut("hooks").and_then(|v| v.as_object_mut()) else {
-        return;
-    };
-
-    for event in CLAUDE_HOOK_EVENTS {
-        let Some(array) = hooks_obj.get_mut(*event).and_then(|v| v.as_array_mut()) else {
-            continue;
-        };
-        array.retain(|matcher_group| !is_paneflow_matcher_group(matcher_group));
-    }
-
-    // Drop empty event arrays.
-    hooks_obj.retain(|_k, v| v.as_array().is_none_or(|a| !a.is_empty()));
-
-    // Drop the empty `hooks` object so the outer file can in turn become
-    // empty and be deleted.
-    if hooks_obj.is_empty() {
-        root_obj.remove("hooks");
-    }
-}
-
-/// Returns `true` if `value` is a matcher-group object that PaneFlow owns.
-/// Belt-and-suspenders: first checks the `_paneflow_managed` marker on the
-/// outer wrapper, then falls back to scanning the inner `hooks` array for a
-/// command whose basename matches `paneflow-ai-hook[.exe]` (legacy bare-name
-/// or absolute-path form). The second pass catches the case where Claude
-/// Code's own settings writer strips unknown fields, AND the case where a
-/// previous shim version wrote a bare-name command (forward compatibility
-/// with the migration to absolute paths).
-fn is_paneflow_matcher_group(value: &serde_json::Value) -> bool {
-    if value
-        .get("_paneflow_managed")
-        .and_then(serde_json::Value::as_bool)
-        .unwrap_or(false)
-    {
-        return true;
-    }
-    value
-        .get("hooks")
-        .and_then(serde_json::Value::as_array)
-        .is_some_and(|inner| {
-            inner.iter().any(|handler| {
-                handler
-                    .get("command")
-                    .and_then(serde_json::Value::as_str)
-                    .is_some_and(is_paneflow_hook_command)
-            })
-        })
-}
-
-/// Atomic write via `NamedTempFile::persist`: create a tempfile in the same
-/// directory (so `persist` can `rename` without crossing filesystems), write
-/// the JSON, then rename over the target. This avoids torn reads if Claude
-/// Code concurrently reads the file at prompt time.
-fn write_atomic(path: &Path, value: &serde_json::Value) -> std::io::Result<()> {
-    let parent = path.parent().ok_or_else(|| {
-        std::io::Error::new(
-            std::io::ErrorKind::InvalidInput,
-            "settings path has no parent",
-        )
-    })?;
-    let mut tmp = tempfile::NamedTempFile::new_in(parent)?;
-    serde_json::to_writer_pretty(&mut tmp, value).map_err(std::io::Error::other)?;
-    // Trailing newline matches what most human editors leave behind and
-    // keeps diffs clean when users inspect the merged file.
-    std::io::Write::write_all(&mut tmp, b"\n")?;
-    tmp.flush()?;
-    tmp.persist(path).map_err(|e| e.error)?;
-    Ok(())
-}
-
-// ---------------------------------------------------------------------------
-// Codex hook config injection (US-006, Unix) —
-//   `.codex/hooks.json` + `~/.codex/config.toml`
-// ---------------------------------------------------------------------------
-
-/// Codex hook events (per Codex docs as of April 2026). Notably `Notification`
-/// is NOT a Codex hook event — Claude Code has one but Codex does not. The
-/// `paneflow-ai-hook` binary (US-003) accepts `Notification` as a valid event
-/// name, but it's only fired from Windows JSONL `error` events (see
-/// `parse_codex_event`), never from this hooks.json registration.
-///
-/// Unix-only: the only callers (`merge_codex_hooks`, `remove_codex_hooks`,
-/// `CodexHookConfigGuard`) are all `#[cfg(unix)]`. On Windows the JSONL tee
-/// path takes over, so this list would otherwise be dead code under
-/// `-D warnings`.
-#[cfg(unix)]
-const CODEX_HOOK_EVENTS: &[&str] = &[
-    "SessionStart",
-    "UserPromptSubmit",
-    "PreToolUse",
-    "PostToolUse",
-    "PermissionRequest",
-    "Stop",
-];
-
-/// Marker for the TOML comment placed above `hooks = true` in
-/// `~/.codex/config.toml`. Cleanup scans for this literal line.
-#[cfg(unix)]
-const CODEX_TOML_MARKER: &str = "# _paneflow_managed: true";
-
-/// Resolve `~/.codex/config.toml` using only std. The shim has no `dirs` dep
-/// and `HOME` is universally set on Unix. Returns `None` on Windows builds
-/// (unreachable at runtime — this function is `#[cfg(unix)]` — but kept for
-/// documentation).
-#[cfg(unix)]
-fn codex_global_config_toml() -> Option<PathBuf> {
-    let home = env::var_os("HOME")?;
-    if home.is_empty() {
-        return None;
-    }
-    Some(PathBuf::from(home).join(".codex").join("config.toml"))
-}
-
-/// RAII guard for Codex's project-level hooks.json + the global config.toml
-/// feature flag. Both are populated on construction and reverted on drop.
-#[cfg(unix)]
-struct CodexHookConfigGuard {
-    hooks_json_path: PathBuf,
-    codex_dir: PathBuf,
-    /// Whether the shim created `./.codex/`. Only rmdir on drop if we did.
-    created_dir: bool,
-    /// Absolute path to `~/.codex/config.toml` (if resolvable); `None` means
-    /// no `HOME` — skip global-config work entirely.
-    config_toml_path: Option<PathBuf>,
-    /// Whether we appended the feature-flag block on install. Drop undoes it
-    /// only if we did; never touches a pre-existing user-authored flag.
-    added_feature_flag: bool,
-}
-
-#[cfg(unix)]
-impl CodexHookConfigGuard {
-    /// Install in the project CWD. Mirrors `HookConfigGuard::install`: same
-    /// orphan-sweep gate when `PANEFLOW_SOCKET_PATH` is unreachable, then
-    /// delegate to `install_at`.
-    fn install() -> Option<Self> {
-        let cwd = env::current_dir().ok()?;
-        let codex_dir = cwd.join(".codex");
-        if !paneflow_ipc_reachable() {
-            sweep_orphan_hook_config(&codex_dir.join("hooks.json"), remove_codex_hooks);
-            return None;
-        }
-        Self::install_at(&codex_dir, codex_global_config_toml().as_deref())
-    }
-
-    /// Testable inner. `config_toml_path` is the absolute path to the global
-    /// `config.toml` (usually `~/.codex/config.toml`); pass `None` to skip
-    /// the feature-flag step entirely (used by tests that don't want to
-    /// pollute the test runner's home dir).
-    fn install_at(codex_dir: &Path, config_toml_path: Option<&Path>) -> Option<Self> {
-        let (hooks_json_path, created_dir) =
-            install_hook_config_file(codex_dir, "hooks.json", "Codex", merge_codex_hooks)?;
-
-        // Codex-specific extra: enable the `hooks = true` feature flag
-        // in `~/.codex/config.toml`. Failure here is non-fatal — the user
-        // can enable it manually and the rest of the hook config is in
-        // place. Runs AFTER the shared install so the per-project hooks.json
-        // is on disk before we touch the global config.
-        let added_feature_flag = config_toml_path
-            .and_then(enable_codex_feature_flag)
-            .unwrap_or(false);
-
-        Some(Self {
-            hooks_json_path,
-            codex_dir: codex_dir.to_path_buf(),
-            created_dir,
-            config_toml_path: config_toml_path.map(Path::to_path_buf),
-            added_feature_flag,
-        })
-    }
-}
-
-#[cfg(unix)]
-impl Drop for CodexHookConfigGuard {
-    fn drop(&mut self) {
-        // Roll back the feature flag first so if something below fails, the
-        // global config is left in a consistent state.
-        if self.added_feature_flag {
-            if let Some(p) = self.config_toml_path.as_deref() {
-                disable_codex_feature_flag(p);
-            }
-        }
-        cleanup_hook_config_file(
-            &self.hooks_json_path,
-            &self.codex_dir,
-            self.created_dir,
-            remove_codex_hooks,
-        );
-    }
-}
-
-#[cfg(unix)]
-fn merge_codex_hooks(root: &mut serde_json::Value) {
-    if !root.is_object() {
-        *root = serde_json::json!({});
-    }
-    let Some(root_obj) = root.as_object_mut() else {
-        return;
-    };
-    let hooks_entry = root_obj
-        .entry("hooks")
-        .or_insert_with(|| serde_json::json!({}));
-    let Some(hooks_obj) = hooks_entry.as_object_mut() else {
-        *hooks_entry = serde_json::json!({});
-        return;
-    };
-
-    for event in CODEX_HOOK_EVENTS {
-        let entry = hooks_obj
-            .entry(*event)
-            .or_insert_with(|| serde_json::json!([]));
-        let Some(array) = entry.as_array_mut() else {
-            continue;
-        };
-        if array.iter().any(is_paneflow_matcher_group) {
-            continue;
-        }
-        array.push(serde_json::json!({
-            "_paneflow_managed": true,
-            "hooks": [
-                {
-                    "type": "command",
-                    "command": resolve_hook_command(event),
-                    "timeout": 5,
-                }
-            ]
-        }));
-    }
-}
-
-#[cfg(unix)]
-fn remove_codex_hooks(root: &mut serde_json::Value) {
-    let Some(root_obj) = root.as_object_mut() else {
-        return;
-    };
-    let Some(hooks_obj) = root_obj.get_mut("hooks").and_then(|v| v.as_object_mut()) else {
-        return;
-    };
-    for event in CODEX_HOOK_EVENTS {
-        let Some(array) = hooks_obj.get_mut(*event).and_then(|v| v.as_array_mut()) else {
-            continue;
-        };
-        array.retain(|m| !is_paneflow_matcher_group(m));
-    }
-    hooks_obj.retain(|_k, v| v.as_array().is_none_or(|a| !a.is_empty()));
-    if hooks_obj.is_empty() {
-        root_obj.remove("hooks");
-    }
-}
-
-/// Append `[features]\nhooks = true` (with a marker comment) to the
-/// file at `path` iff: (a) the file doesn't already have `hooks = true` (or
-/// the pre-Codex-0.130 alias `codex_hooks = true`) anywhere, AND (b) there's
-/// no existing `[features]` section — appending would create a duplicate-
-/// section TOML error, so we abstain in that case and warn.
-///
-/// Returns `Some(true)` if we modified the file, `Some(false)` if the flag
-/// was already present (no-op), `None` if we aborted due to a conflict or
-/// I/O error.
-/// US-027: RAII advisory `flock` guard. Serializes the codex `[features]`
-/// read-modify-write across concurrent shims so two near-simultaneous `codex`
-/// launches can't both append `[features]` (→ duplicate-section invalid TOML).
-#[cfg(unix)]
-struct FlockGuard {
-    file: std::fs::File,
-}
-
-#[cfg(unix)]
-impl FlockGuard {
-    /// Acquire an exclusive advisory lock on `lock_path` (creating it),
-    /// blocking until granted. Returns `None` if the lock file can't be opened
-    /// or locked — callers then proceed best-effort (no worse than no lock).
-    fn acquire(lock_path: &Path) -> Option<Self> {
-        use std::os::unix::io::AsRawFd;
-        let file = std::fs::OpenOptions::new()
-            .read(true)
-            .write(true)
-            .create(true)
-            .truncate(false)
-            .open(lock_path)
-            .ok()?;
-        // SAFETY: `flock` on a valid owned fd; `LOCK_EX` blocks until the lock
-        // is exclusively held. The lock is released in `Drop`.
-        let rc = unsafe { libc::flock(file.as_raw_fd(), libc::LOCK_EX) };
-        (rc == 0).then_some(Self { file })
-    }
-}
-
-#[cfg(unix)]
-impl Drop for FlockGuard {
-    fn drop(&mut self) {
-        use std::os::unix::io::AsRawFd;
-        // SAFETY: releasing our own advisory lock; ignore teardown errors.
-        unsafe {
-            libc::flock(self.file.as_raw_fd(), libc::LOCK_UN);
-        }
-    }
-}
-
-#[cfg(unix)]
-fn enable_codex_feature_flag(path: &Path) -> Option<bool> {
-    // US-027: hold an exclusive flock across the whole read-modify-write.
-    // `write_text_atomic` guarantees byte-atomicity but NOT serialization, so
-    // without this two concurrent shims both read a config lacking
-    // `[features]`, both pass the guard below, and both append the section.
-    // A failed lock acquisition degrades to the prior (unserialized) behavior.
-    let _lock = path.parent().and_then(|dir| {
-        let _ = std::fs::create_dir_all(dir);
-        FlockGuard::acquire(&dir.join(".paneflow-codex.lock"))
-    });
-
-    let existing = match std::fs::read_to_string(path) {
-        Ok(s) => s,
-        Err(e) if e.kind() == std::io::ErrorKind::NotFound => String::new(),
-        Err(e) => {
-            eprintln!(
-                "paneflow-shim: cannot read {} ({e}); leaving Codex `hooks` feature flag alone",
-                safe_path_display(path)
-            );
-            return None;
-        }
-    };
-
-    if has_hooks_flag(&existing) {
-        return Some(false);
-    }
-    if has_features_section(&existing) {
-        eprintln!(
-            "paneflow-shim: {} already has a [features] section without `hooks`; skipping auto-enable (add `hooks = true` there manually to enable Codex hooks)",
-            safe_path_display(path)
-        );
-        return None;
-    }
-
-    // Safe append: ensure a trailing newline, then add our block.
-    let mut next = existing.clone();
-    if !next.is_empty() && !next.ends_with('\n') {
-        next.push('\n');
-    }
-    if !next.is_empty() {
-        next.push('\n');
-    }
-    next.push_str(CODEX_TOML_MARKER);
-    next.push('\n');
-    next.push_str("[features]\nhooks = true\n");
-
-    // Ensure the parent dir exists — the Codex config dir may be absent
-    // if the user has never run Codex before, but the shim's own invocation
-    // implies the binary is installed somewhere.
-    if let Some(parent) = path.parent() {
-        let _ = std::fs::create_dir_all(parent);
-    }
-
-    if let Err(e) = write_text_atomic(path, &next) {
-        eprintln!(
-            "paneflow-shim: cannot write {} ({e}); Codex `hooks` feature flag not set",
-            safe_path_display(path)
-        );
-        return None;
-    }
-    Some(true)
-}
-
-#[cfg(unix)]
-fn disable_codex_feature_flag(path: &Path) {
-    let Ok(existing) = std::fs::read_to_string(path) else {
-        return;
-    };
-    let Some(cleaned) = strip_codex_feature_block(&existing) else {
-        // Marker not found, or the managed block was edited by the user.
-        // Both cases: don't touch the file. No stderr noise — next shim
-        // invocation's idempotent install will converge state eventually.
-        return;
-    };
-    // If the file is now empty or only whitespace, delete it — we either
-    // created it from nothing or the flag was the only content.
-    let result = if cleaned.trim().is_empty() {
-        std::fs::remove_file(path)
-    } else {
-        write_text_atomic(path, &cleaned)
-    };
-    if let Err(e) = result {
-        // Phase 7 security audit MEDIUM #12: if cleanup write fails, the
-        // managed block stays in `~/.codex/config.toml` and subsequent
-        // shim runs see `hooks = true` already set, so they never
-        // retry the cleanup. Surface the failure so the user can remove
-        // the block manually. Known silent-failure mode.
-        eprintln!(
-            "paneflow-shim: could not revert {} ({e}); please remove the `{}` block manually",
-            safe_path_display(path),
-            CODEX_TOML_MARKER
-        );
-    }
-}
-
-/// Atomic text write via `tempfile::NamedTempFile::persist`. Mirrors the
-/// existing `write_atomic` but for raw strings (not serde_json::Value).
-/// Phase 7 security audit MEDIUM #3: concurrent shims both calling
-/// `fs::write` on `~/.codex/config.toml` can produce torn bytes; rename
-/// swap is the fix.
-#[cfg(unix)]
-fn write_text_atomic(path: &Path, content: &str) -> std::io::Result<()> {
-    let parent = path.parent().ok_or_else(|| {
-        std::io::Error::new(
-            std::io::ErrorKind::InvalidInput,
-            "config path has no parent",
-        )
-    })?;
-    let mut tmp = tempfile::NamedTempFile::new_in(parent)?;
-    std::io::Write::write_all(&mut tmp, content.as_bytes())?;
-    tmp.flush()?;
-    tmp.persist(path).map_err(|e| e.error)?;
-    Ok(())
-}
-
-#[cfg(unix)]
-fn has_hooks_flag(content: &str) -> bool {
-    // Codex 0.130 renamed `codex_hooks` -> `hooks`; accept both so the shim
-    // stays silent for configs from either era.
-    content.lines().any(|line| {
-        let l = line.trim_start();
-        if l.starts_with('#') {
-            return false;
-        }
-        let stripped = l.split_once('#').map(|(k, _)| k).unwrap_or(l);
-        let stripped = stripped.trim_end();
-        match stripped.split_once('=') {
-            Some((key, value)) => {
-                let k = key.trim();
-                (k == "hooks" || k == "codex_hooks") && value.trim() == "true"
-            }
-            None => false,
-        }
-    })
-}
-
-#[cfg(unix)]
-fn has_features_section(content: &str) -> bool {
-    content.lines().any(|line| line.trim() == "[features]")
-}
-
-/// Remove the exact 3-line PaneFlow block (marker comment + `[features]` +
-/// `hooks = true`, or the legacy `codex_hooks = true` for installs predating
-/// Codex 0.130) from the file content. Returns the cleaned content if the
-/// block was found and removed; `None` if the marker wasn't present.
-#[cfg(unix)]
-fn strip_codex_feature_block(content: &str) -> Option<String> {
-    let lines: Vec<&str> = content.lines().collect();
-    let marker_idx = lines.iter().position(|l| l.trim() == CODEX_TOML_MARKER)?;
-    // We expect the marker to be followed by exactly:
-    //   `[features]`
-    //   `hooks = true` (or the legacy `codex_hooks = true`)
-    // Remove those three lines; anything else means the file was edited
-    // and we should bail to avoid clobbering user content.
-    let tail = lines.get(marker_idx + 1..marker_idx + 3)?;
-    // Exact-match both lines: `starts_with("hooks")` would also strip a
-    // hypothetical future `hooks_experimental = ...` line that happened
-    // to sit in the managed block. Exact match fails closed — safer to
-    // leave the block untouched than over-delete. Accept either key name
-    // because installs predating the Codex 0.130 rename wrote `codex_hooks`.
-    let second = tail[1].trim();
-    if tail[0].trim() != "[features]"
-        || (second != "hooks = true" && second != "codex_hooks = true")
-    {
-        return None;
-    }
-
-    // Reconstruct: before marker + (possibly one blank line preceding the
-    // marker that we added) + after the 3-line block.
-    let mut head_end = marker_idx;
-    // Strip a single trailing blank line we may have added before the marker
-    // so we don't accumulate blanks across install/uninstall cycles.
-    if head_end > 0 && lines[head_end - 1].is_empty() {
-        head_end -= 1;
-    }
-
-    let mut out = String::new();
-    for line in &lines[..head_end] {
-        out.push_str(line);
-        out.push('\n');
-    }
-    for line in &lines[marker_idx + 3..] {
-        out.push_str(line);
-        out.push('\n');
-    }
-    // Preserve original EOF-newline behavior: if the input had no trailing
-    // newline and we removed content from the tail, trim one.
-    if !content.ends_with('\n') && out.ends_with('\n') {
-        out.pop();
-    }
-    Some(out)
-}
-
-// ---------------------------------------------------------------------------
-// Codex Windows JSONL fallback (US-006, non-Unix)
-// ---------------------------------------------------------------------------
-//
-// Codex hook machinery is disabled on Windows as of April 2026. Instead, the
-// shim spawns `codex exec --json`, tees the child's stdout (so the user sees
-// every byte unchanged), and dispatches `paneflow-ai-hook <Event>` subprocesses
-// on recognized NDJSON events. Only the `exec` subcommand supports `--json`;
-// interactive `codex` invocations on Windows pass through without tee — they
-// simply won't produce sidebar state updates (documented regression).
-
-#[cfg(not(unix))]
-/// Decide whether to tee + inject `--json`. The `exec` subcommand is the only
-/// Codex mode that emits NDJSON; other invocations (`codex`, `codex resume`,
-/// `codex --help`) must pass through unmodified.
-///
-/// Scans for `"exec"` anywhere in argv rather than only at position 0, so
-/// that global flags preceding the subcommand (e.g.,
-/// `codex --config cfg.toml exec prompt`) are still detected. Known false-
-/// positive edge: a user who passes `exec` as the VALUE of a preceding
-/// value-taking flag (e.g., `codex --profile exec`) would get `--json`
-/// injected; Codex then errors clearly and the user can bypass the shim
-/// by invoking the real binary directly. Accepting this trade because
-/// missing the tee on `codex --config X exec` is a silent regression,
-/// while the false-positive is loud and bypassable.
-fn rewrite_codex_args(args: &[OsString]) -> (Vec<OsString>, bool) {
-    let Some(exec_idx) = args.iter().position(|a| a == "exec") else {
-        return (args.to_vec(), false);
-    };
-    let mut rewritten = Vec::with_capacity(args.len() + 1);
-    rewritten.extend_from_slice(&args[..=exec_idx]);
-    rewritten.push(OsString::from("--json"));
-    rewritten.extend(args[exec_idx + 1..].iter().cloned());
-    (rewritten, true)
-}
-
-/// The full catalog of Codex NDJSON `"type"` discriminators documented at
-/// <https://developers.openai.com/codex/noninteractive> as of April 2026.
-/// `parse_codex_event`'s match arms MUST exhaustively cover every string in
-/// this list so the schema-pin test can detect drift. When Codex adds a new
-/// event type, add it to BOTH this const AND to a `match` arm in
-/// `parse_codex_event` (even if the arm is just `None`).
-///
-/// Test-only: the runtime `parse_codex_event` enumerates the strings inline
-/// rather than referencing this const, so without `cfg(test)` the const is
-/// dead code under `-D warnings` on Windows non-test builds.
-#[cfg(all(test, not(unix)))]
-const KNOWN_CODEX_EVENT_TYPES: &[&str] = &[
-    "thread.started",
-    "turn.started",
-    "turn.completed",
-    "turn.failed",
-    "item.started",
-    "item.completed",
-    "error",
-];
-
-/// Map a Codex NDJSON line to a `paneflow-ai-hook` event name. Returns
-/// `None` for unrecognized / sub-event / malformed lines so the caller can
-/// log-and-skip without breaking the tee loop.
-#[cfg(not(unix))]
-fn parse_codex_event(line: &str) -> Option<&'static str> {
-    let trimmed = line.trim();
-    if trimmed.is_empty() {
-        return None;
-    }
-    let value: serde_json::Value = serde_json::from_str(trimmed).ok()?;
-    let event_type = value.get("type")?.as_str()?;
-    match event_type {
-        "turn.started" => Some("UserPromptSubmit"),
-        "turn.completed" => Some("Stop"),
-        "error" => Some("Notification"),
-        // Known-but-unmapped events: explicit arms (not a catch-all) so
-        // adding a new Codex event type without touching this match arm
-        // requires a code change, which the schema-pin test will then
-        // either accept (if the fixture is updated) or fail loudly.
-        "thread.started" | "turn.failed" | "item.started" | "item.completed" => None,
-        // Truly unknown (new Codex event type we don't recognize yet):
-        // silently skip. The schema-pin test cross-checks this against
-        // KNOWN_CODEX_EVENT_TYPES so drift is surfaced at dev time.
-        _ => None,
-    }
-}
-
-#[cfg(not(unix))]
-fn run_codex_with_jsonl_tee(path: &Path, args: &[OsString]) -> ExitCode {
-    use std::io::{BufRead, BufReader};
-    use std::process::Stdio;
-
-    let mut child = match std::process::Command::new(path)
-        .args(args)
-        .envs(env::vars_os())
-        .env("PANEFLOW_AI_TOOL", "codex")
-        .env("PANEFLOW_AI_PID", std::process::id().to_string())
-        .stdin(Stdio::inherit())
-        .stdout(Stdio::piped())
-        .stderr(Stdio::inherit())
-        .spawn()
-    {
-        Ok(c) => c,
-        Err(e) => {
-            eprintln!(
-                "paneflow-shim: spawn '{}' failed: {e}",
-                safe_path_display(path)
-            );
-            return ExitCode::from(127);
-        }
-    };
-
-    // Resolve the sibling hook binary ONCE before the tee loop. Bare-name
-    // `Command::new("paneflow-ai-hook")` works on Windows only when the
-    // shim cache dir is already on `%PATH%`, which is a per-PTY contract
-    // not guaranteed at the moment this function spawns its first hook.
-    // Resolving by absolute path matches every other call site in this
-    // crate (`send_interrupt_stop`, `notify_session_end`).
-    let hook_path = locate_sibling_hook_binary();
-
-    // Take stdout so we can tee it. If the child was spawned without piped
-    // stdout somehow, skip the tee entirely and fall through to wait().
-    if let Some(stdout) = child.stdout.take() {
-        // Reader thread: keeps the child's stdout drained (prevents pipe
-        // fills) and dispatches hook events as they arrive.
-        let tee_handle = std::thread::spawn(move || {
-            let mut out = std::io::stdout().lock();
-            let reader = BufReader::new(stdout);
-            for line in reader.lines().map_while(Result::ok) {
-                // Echo the line verbatim, preserving Codex's NDJSON so any
-                // downstream tooling (jq, log-tailers) still works.
-                let _ = std::io::Write::write_all(&mut out, line.as_bytes());
-                let _ = std::io::Write::write_all(&mut out, b"\n");
-                let _ = std::io::Write::flush(&mut out);
-
-                if let Some(event) = parse_codex_event(&line) {
-                    // Fire-and-forget: we don't wait on the hook binary so
-                    // the tee loop keeps pace with Codex's output. This
-                    // function is compiled only on non-Unix (`#[cfg(not
-                    // (unix))]`) — on Windows, dropping the `Child` handle
-                    // releases the OS handle and the subprocess runs to
-                    // completion independently, with the OS reaping it
-                    // when it exits. No zombies, no cleanup required.
-                    //
-                    // Known limitation (Phase 7 audit, MEDIUM #8): there is
-                    // no rate limit. If Codex ever emits thousands of events
-                    // per second, the shim would spawn thousands of
-                    // subprocesses per second. Typical Codex output is a
-                    // few events/sec, so this is tolerated without state.
-                    if let Some(ref hp) = hook_path {
-                        let _ = std::process::Command::new(hp)
-                            .arg(event)
-                            .stdin(Stdio::null())
-                            .stdout(Stdio::null())
-                            .stderr(Stdio::null())
-                            .spawn();
-                    }
-                }
-            }
-        });
-        // Join the reader thread AFTER wait() so all buffered output makes
-        // it to the user before we exit.
-        let status = child.wait();
-        let _ = tee_handle.join();
-        match status {
-            Ok(s) => exit_code_from_status(&s),
-            Err(e) => {
-                eprintln!(
-                    "paneflow-shim: wait on '{}' failed: {e}",
-                    safe_path_display(path)
-                );
-                ExitCode::from(127)
-            }
-        }
-    } else {
-        match child.wait() {
-            Ok(s) => exit_code_from_status(&s),
-            Err(e) => {
-                eprintln!(
-                    "paneflow-shim: wait on '{}' failed: {e}",
-                    safe_path_display(path)
-                );
-                ExitCode::from(127)
-            }
-        }
-    }
-}
+mod detect;
+mod exec;
+mod hooks;
+
+use detect::*;
+use exec::*;
+use hooks::*;
 
 // ---------------------------------------------------------------------------
 // Orchestrator
@@ -1639,7 +143,7 @@ fn notify_session_end(tool: &str) {
 /// directory as this shim binary. Returns `None` if `current_exe()`
 /// fails or the sibling isn't a regular file — in either case, the
 /// caller silently skips notification.
-fn locate_sibling_hook_binary() -> Option<PathBuf> {
+pub(crate) fn locate_sibling_hook_binary() -> Option<PathBuf> {
     let exe = env::current_exe().ok()?;
     let dir = exe.parent()?;
     #[cfg(unix)]
@@ -1657,6 +161,7 @@ fn locate_sibling_hook_binary() -> Option<PathBuf> {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use std::path::Path;
 
     #[test]
     fn detect_tool_from_stem_maps_known_stems() {
